refactor: rename trait to concern
