+++ src/Commands/Server/Certificate/CreateCaCert.php
@@ -6,11 +6,11 @@ namespace Turso\Driver\Laravel\Commands\Server\Certificate;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Process;
-use Turso\Driver\Laravel\Traits\CommandTrait;
+use Turso\Driver\Laravel\Concerns\HandlesTursoInstallerCommands;
 
 class CreateCaCert extends Command
 {
-    use CommandTrait;
+    use HandlesTursoInstallerCommands;
 
     protected $signature = 'turso-php:ca-cert-create {name=ca}
         {--expiry=30 : Expiry in days, default is 30 days}
@@ -40,8 +40,6 @@ class CreateCaCert extends Command
 
         if ($process->failed()) {
             $this->error('Failed to create CA certificate.');
-
-            return;
         }
     }
 }
