+++ src/Commands/Server/Certificate/ShowCaCert.php
@@ -6,11 +6,11 @@ namespace Turso\Driver\Laravel\Commands\Server\Certificate;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Process;
-use Turso\Driver\Laravel\Traits\CommandTrait;
+use Turso\Driver\Laravel\Concerns\HandlesTursoInstallerCommands;
 
 class ShowCaCert extends Command
 {
-    use CommandTrait;
+    use HandlesTursoInstallerCommands;
 
     protected $signature = 'turso-php:ca-cert-show
         {--raw : Show raw CA certificate and private key}
