+++ src/Commands/Server/CheckRequirementServer.php
@@ -6,11 +6,11 @@ namespace Turso\Driver\Laravel\Commands\Server;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Process;
-use Turso\Driver\Laravel\Traits\CommandTrait;
+use Turso\Driver\Laravel\Concerns\HandlesTursoInstallerCommands;
 
 class CheckRequirementServer extends Command
 {
-    use CommandTrait;
+    use HandlesTursoInstallerCommands;
 
     protected $signature = 'turso-php:server-check';
