+++ src/Commands/Sqld/Environment/CreateEnvironment.php
@@ -6,11 +6,11 @@ namespace Turso\Driver\Laravel\Commands\Sqld\Environment;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Process;
-use Turso\Driver\Laravel\Traits\CommandTrait;
+use Turso\Driver\Laravel\Concerns\HandlesTursoInstallerCommands;
 
 class CreateEnvironment extends Command
 {
-    use CommandTrait;
+    use HandlesTursoInstallerCommands;
 
     protected $signature = 'turso-php:env-create {name : The name of the environment}
         {--variables= : The variables of the environment in JSON/DSN format}
