+++ src/Commands/Token/ListToken.php
@@ -6,11 +6,11 @@ namespace Turso\Driver\Laravel\Commands\Token;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Process;
-use Turso\Driver\Laravel\Traits\CommandTrait;
+use Turso\Driver\Laravel\Concerns\HandlesTursoInstallerCommands;
 
 class ListToken extends Command
 {
-    use CommandTrait;
+    use HandlesTursoInstallerCommands;
 
     protected $signature = 'turso-php:token-list';
